Create Binary_To_Eight
First instantiated the module "ssd" that creates the instance for the seven segment display on the FPGA and what the inputs would show on the display. Then, the SimpleALU module is instantiated to run certain operations between the inputs of the FPGA. The seven segment display module is called upon in SimpleALU to use the result of the operations for the seven segment display input.