Create 4_Bit_Counter
Instantiated a Counter module that simulates 4 toggle flip-flops to create a counter. A seven segment display module is utilized to show the value of the counter as the counter progresses. Reset allows the counter to reset once it reaches a certain value, and select controls the way the counter operates, whether or not the counter counts from high to low or low to high.